fix: wezterm open dir
