+++ FinderSyncExt/Info.plist
@@ -2,6 +2,7 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+
 	<key>NSExtension</key>
 	<dict>
 		<key>NSExtensionAttributes</key>
