+++ RClick/MenuBarView.swift
@@ -9,13 +9,18 @@ import AppKit
 import SwiftUI
 
 struct MenuBarView: View {
+    @Environment(\.openWindow) var openWindow: OpenWindowAction
+    
     let messager = Messager.shared
+    
     var body: some View {
         VStack {
-            SettingsLink {
+           
+            Button(action: actionSettings) {
                 Image(systemName: "gearshape")
                 Text("Settings")
-            }.keyboardShortcut(",", modifiers: [.command])
+            }
+            .keyboardShortcut(",", modifiers: [.command])
             
             Button(action: actionQuit) {
                 Image(systemName: "xmark.square")
@@ -26,8 +31,8 @@ struct MenuBarView: View {
         }
     }
 
-    func test () {
-        
+    private func actionSettings () {
+        openWindow(id: Constants.settingsWindowID)
     }
     private func actionQuit() {
