+++ RClick/RClickApp.swift
@@ -9,6 +9,12 @@ import Foundation
 import SwiftData
 import SwiftUI
 
+import FinderSync
+import os.log
+
+
+private let logger = Logger(subsystem: subsystem, category: "AppDelegate")
+
 let channel = AppCommChannel()
 
 @main
@@ -23,10 +29,8 @@ struct RClickApp: App {
     private var logger
 
     var body: some Scene {
-        Settings {
-            SettingsView()
-        }
-        .defaultAppStorage(.group)
+        SettingsWindow(onAppear: {})
+            .defaultAppStorage(.group)
 
         MenuBarExtra(
             "RClick", image: "MenuBar", isInserted: self.$showMenuBarExtra
@@ -36,126 +40,44 @@ struct RClickApp: App {
     }
 }
 
-// https://stackoverflow.com/a/76714125/19625526
-private let kAppMenuInternalIdentifier = "app"
-private let kSettingsLocalizedStringKey = "Settings\\U2026"
 
-extension NSApplication {
-    /// Open the application settings/preferences window.
-    func openSettings() {
-        // macOS 14 Sonoma
-        if let internalItemAction = NSApp.mainMenu?.item(
-            withInternalIdentifier: kAppMenuInternalIdentifier
-        )?.submenu?.item(
-            withLocalizedTitle: kSettingsLocalizedStringKey
-        )?.internalItemAction {
-            internalItemAction()
-            return
+
+class AppDelegate: NSObject, NSApplicationDelegate {
+    let messager = Messager.shared
+    var folderItemStore = FolderItemStore()
+    var showDockIcon = UserDefaults.group.bool(forKey: Key.showDockIcon)
+    var settingsWindow: NSWindow!
+
+    func applicationDidFinishLaunching(_ aNotification: Notification) {
+        // 在 app 启动后执行的函数
+        logger.notice("App -------------------------- 已启动")
+
+        Task {
+            await channel.setup(store: folderItemStore)
         }
 
-        guard let delegate = NSApp.delegate else { return }
+        messager.start(name: Key.messageFromFinder)
+        messager.sendMessage(name: "running", data: MessagePayload(action: "running"))
 
-        // macOS 13 Ventura
-        var selector = Selector(("showSettingsWindow:"))
-        if delegate.responds(to: selector) {
-            delegate.perform(selector, with: nil, with: nil)
-            return
-        }
+//        // 根据某种逻辑设置应用是否显示在 Dock 中
+//        if showDockIcon {
+//            NSApp.setActivationPolicy(.regular)
+//        } else {
+//            NSApp.setActivationPolicy(.prohibited)
+//        }
+    }
 
-        // macOS 12 Monterrey
-        selector = Selector(("showPreferencesWindow:"))
-        if delegate.responds(to: selector) {
-            delegate.perform(selector, with: nil, with: nil)
-            return
-        }
+//    func applicationDidBecomeActive(_ notification: Notification) {
+//        NSApplication.shared.openSettings()
+//    }
+
+    func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
+        return false
+    }
+
+    func applicationWillTerminate(_ notification: Notification) {
+        messager.sendMessage(name: "quit", data: MessagePayload(action: "quit"))
+        logger.info("applicationWillTerminate")
     }
 }
 
-// MARK: - NSMenuItem (Private)
-
-extension NSMenuItem {
-    /// An internal SwiftUI menu item identifier that should be a public property on `NSMenuItem`.
-    var internalIdentifier: String? {
-        guard let id = Mirror.firstChild(
-            withLabel: "id", in: self
-        )?.value else {
-            return nil
-        }
-
-        return "\(id)"
-    }
-
-    /// A callback which is associated directly with this `NSMenuItem`.
-    var internalItemAction: (() -> Void)? {
-        guard
-            let platformItemAction = Mirror.firstChild(
-                withLabel: "platformItemAction", in: self
-            )?.value,
-            let typeErasedCallback = Mirror.firstChild(
-                in: platformItemAction)?.value
-        else {
-            return nil
-        }
-
-        return Mirror.firstChild(
-            in: typeErasedCallback
-        )?.value as? () -> Void
-    }
-}
-
-// MARK: - NSMenu (Private)
-
-extension NSMenu {
-    /// Get the first `NSMenuItem` whose internal identifier string matches the given value.
-    func item(withInternalIdentifier identifier: String) -> NSMenuItem? {
-        self.items.first(where: {
-            $0.internalIdentifier?.elementsEqual(identifier) ?? false
-        })
-    }
-
-    /// Get the first `NSMenuItem` whose title is equivalent to the localized string referenced
-    /// by the given localized string key in the localization table identified by the given table name
-    /// from the bundle located at the given bundle path.
-    func item(
-        withLocalizedTitle localizedTitleKey: String,
-        inTable tableName: String = "MenuCommands",
-        fromBundle bundlePath: String = "/System/Library/Frameworks/AppKit.framework"
-    ) -> NSMenuItem? {
-        guard let localizationResource = Bundle(path: bundlePath) else {
-            return nil
-        }
-
-        return self.item(withTitle: NSLocalizedString(
-            localizedTitleKey,
-            tableName: tableName,
-            bundle: localizationResource,
-            comment: ""
-        ))
-    }
-}
-
-// MARK: - Mirror (Helper)
-
-private extension Mirror {
-    /// The unconditional first child of the reflection subject.
-    var firstChild: Child? { self.children.first }
-
-    /// The first child of the reflection subject whose label matches the given string.
-    func firstChild(withLabel label: String) -> Child? {
-        self.children.first(where: {
-            $0.label?.elementsEqual(label) ?? false
-        })
-    }
-
-    /// The unconditional first child of the given subject.
-    static func firstChild(in subject: Any) -> Child? {
-        Mirror(reflecting: subject).firstChild
-    }
-
-    /// The first child of the given subject whose label matches the given string.
-    static func firstChild(
-        withLabel label: String, in subject: Any
-    ) -> Child? {
-        Mirror(reflecting: subject).firstChild(withLabel: label)
-    }
-}
