+++ RClick/Settings/AboutSettingsTabView.swift
