+++ RClick/Shared/Constants.swift
@@ -1,0 +1,16 @@
+//
+//  Constants.swift
+//  RClick
+//
+//  Created by 李旭 on 2024/9/25.
+//
+
+import Foundation
+
+
+enum Constants {
+   
+    /// The identifier for the settings window.
+    static let settingsWindowID = "rclick-settings"
+
+}
